[REFACTOR] 함수를 각 파일로 분리하여 정리
Fixes: TICO-40
Related to: TICO-33
