+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/timer/view/CustomCircularDraggableTimer.kt
@@ -4,18 +4,9 @@ import android.graphics.Paint
 import android.graphics.Typeface
 import android.util.Log
 import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectDragGestures
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -40,195 +31,14 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.res.ResourcesCompat
 import com.tico.pomorodo.R
-import com.tico.pomorodo.ui.common.view.CustomTextButton
-import com.tico.pomorodo.ui.common.view.CustomTimeText
 import com.tico.pomorodo.ui.iconpack.commonIconPack.IcTimerPin
 import com.tico.pomorodo.ui.theme.IconPack
-import com.tico.pomorodo.ui.theme.PomoroDoTheme
 import kotlin.math.PI
 import kotlin.math.atan2
 import kotlin.math.cos
 import kotlin.math.hypot
 import kotlin.math.roundToInt
 import kotlin.math.sin
-
-@Composable
-fun TimerRootScreen() {
-    var concentrationTime by remember {
-        mutableIntStateOf(30)
-    }
-    var breakTime by remember {
-        mutableIntStateOf(70)
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 30.dp, vertical = 24.dp),
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        PomodoroTimerScreen(
-            concentrationTime = concentrationTime,
-            breakTime = breakTime,
-            onConcentrationTimeChange = { position -> concentrationTime = position },
-            onBreakTimeChange = { position -> breakTime = position }
-        )
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        CustomTextButton(
-            text = stringResource(R.string.content_button_start),
-            backgroundColor = PomoroDoTheme.colorScheme.primaryContainer,
-            textColor = PomoroDoTheme.colorScheme.background,
-            textStyle = PomoroDoTheme.typography.laundryGothicRegular18,
-            verticalPadding = 12.dp
-        ) { /*TODO*/ }
-    }
-}
-
-@Composable
-fun PomodoroTimerScreen(
-    concentrationTime: Int,
-    breakTime: Int,
-    onConcentrationTimeChange: (Int) -> Unit,
-    onBreakTimeChange: (Int) -> Unit
-) {
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        CustomPomodoroTimer(
-            concentrationTime = concentrationTime,
-            breakTime = breakTime,
-            onConcentrationTimeChange = onConcentrationTimeChange,
-            onBreakTimeChange = onBreakTimeChange
-        )
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        TodayConcentrationInformation()
-    }
-}
-
-@Composable
-fun TodayConcentrationInformation() {
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        CustomTimeText(
-            title = stringResource(R.string.title_taget_concentration),
-            hour = 12,
-            minute = 0,
-            second = 0,
-            textColor = PomoroDoTheme.colorScheme.onBackground,
-            spaceDp = 4.dp,
-            textStyle = PomoroDoTheme.typography.laundryGothicRegular18,
-        )
-
-        Spacer(modifier = Modifier.height(14.dp))
-
-        Row(
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            CustomTimeText(
-                title = stringResource(R.string.title_total_concentration),
-                hour = 10,
-                minute = 0,
-                second = 0,
-                textColor = PomoroDoTheme.colorScheme.onBackground,
-                spaceDp = 4.dp,
-                textStyle = PomoroDoTheme.typography.laundryGothicRegular18,
-            )
-
-            Spacer(modifier = Modifier.width(50.dp))
-
-            CustomTimeText(
-                title = stringResource(R.string.title_total_break),
-                hour = 3,
-                minute = 0,
-                second = 0,
-                textColor = PomoroDoTheme.colorScheme.onBackground,
-                spaceDp = 4.dp,
-                textStyle = PomoroDoTheme.typography.laundryGothicRegular18,
-            )
-        }
-    }
-}
-
-@Composable
-fun CustomPomodoroTimer(
-    concentrationTime: Int,
-    breakTime: Int,
-    onConcentrationTimeChange: (Int) -> Unit,
-    onBreakTimeChange: (Int) -> Unit,
-) {
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        EditableTextTimer(
-            concentrationTime = concentrationTime,
-            breakTime = breakTime,
-            concentrationColor = PomoroDoTheme.colorScheme.primaryContainer,
-            breakColor = PomoroDoTheme.colorScheme.secondaryContainer
-        )
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        CustomCircularDraggableTimer(
-            modifier = Modifier
-                .size(300.dp)
-                .background(PomoroDoTheme.colorScheme.background),
-            outerTimerColor = PomoroDoTheme.colorScheme.primaryContainer,
-            innerTimerColor = PomoroDoTheme.colorScheme.secondaryContainer,
-            backgroundColor = PomoroDoTheme.colorScheme.background,
-            indicatorColor = PomoroDoTheme.colorScheme.onBackground,
-            outerCircleRadius = 125,
-            outerInitialValue = concentrationTime,
-            innerInitialValue = breakTime,
-            onOuterPositionChange = { position -> onConcentrationTimeChange(position) },
-            onInnerPositionChange = { position -> onBreakTimeChange(position) }
-        )
-    }
-}
-
-@Composable
-fun EditableTextTimer(
-    concentrationTime: Int,
-    breakTime: Int,
-    concentrationColor: Color,
-    breakColor: Color
-) {
-    Row(
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        CustomTimeText(
-            title = stringResource(R.string.title_concentration_time),
-            hour = concentrationTime / 60,
-            minute = concentrationTime % 60,
-            textColor = concentrationColor,
-            spaceDp = 10.dp,
-            titleTextStyle = PomoroDoTheme.typography.laundryGothicRegular18,
-            contentTextStyle = PomoroDoTheme.typography.laundryGothicRegular16
-        )
-
-        Spacer(modifier = Modifier.width(20.dp))
-
-        CustomTimeText(
-            title = stringResource(R.string.title_break_time),
-            hour = breakTime / 60,
-            minute = breakTime % 60,
-            textColor = breakColor,
-            spaceDp = 10.dp,
-            titleTextStyle = PomoroDoTheme.typography.laundryGothicRegular18,
-            contentTextStyle = PomoroDoTheme.typography.laundryGothicRegular16
-        )
-    }
-}
 
 @Composable
 fun CustomCircularDraggableTimer(
